requirements.txt
ollama
faster-whisper
piper-tts
opencv-python
Pillow
face-recognition
pyaudio
numpy
